+++ casbah-query/src/main/scala/CoreOperators.scala
@@ -128,7 +128,7 @@ trait QueryOperator extends ChainedOperator {
    * WARNING: This does NOT check that target is a serializable type.
    * That is, for the moment, your own problem.
    */
-  protected def queryEq(target: Any): DBObject with QueryExpressionObject = QueryExpressionObject((field -> target))
+  protected def queryEq(target: Any): DBObject with QueryExpressionObject = QueryExpressionObject(field -> target)
 
 }
 
@@ -435,9 +435,29 @@ trait TypeOp extends QueryOperator {
   def $type[A](implicit bsonType: BSONType[A]) = queryOp(oper, bsonType.operator)
 }
 
+/**
+ * \$regex operator to query by type.
+ *
+ * Takes a string for use in the $regex query
+ *
+ * "foo" $regex "^bar$"
+ *
+ * @since 2.6.2
+ * @see http://www.mongodb.org/display/DOCS/Advanced+Queries#AdvancedQueries-%7B%7B%24type%7D%7D
+ */
+trait regexOp extends QueryOperator {
+  private val oper = "$regex"
+
+  def $regex(arg: String) = queryOp(oper, arg)
+}
+
+// Geo Spatial Ops
+
 trait GeospatialOps extends GeoNearOp
   with GeoNearSphereOp
   with GeoWithinOps
+  with GeoIntersectsOp
+  with DeprecatedGeoWithinOps
 
 case class GeoCoords[A: ValidNumericType: Manifest, B: ValidNumericType: Manifest](val lat: A, val lon: B) {
   def toList = MongoDBList(lat, lon)
@@ -494,6 +514,67 @@ trait GeoNearSphereOp extends QueryOperator {
 
 /**
  *
+ * Trait to provide the nested \$geoWithin geospatial search method on appropriate callers
+ *
+ *
+ * Note that  the args aren't TECHNICALLY latitude and longitude as they depend on:
+ *   a) the order you specified your actual index in
+ *   b) if you're using actual world maps or something else
+ *
+ * @since 2.6.2
+ * @see http://www.mongodb.org/display/DOCS/Geospatial+Indexing
+ */
+trait GeoWithinOps extends QueryOperator {
+  self =>
+  private val oper = "$geoWithin"
+
+  def $geoWithin[A <% DBObject](geometry: A) = queryOp(oper, geometry)
+  def $geoWithin = new QueryOperator {
+    val field = "$geoWithin"
+
+    def $polygon(coords: Iterable[GeoCoords[_, _]]) =
+      MongoDBObject(
+        self.field ->
+        queryOp("$polygon", coords.toList)
+      )
+    def $polygon(coords: GeoCoords[_, _]*) =
+      MongoDBObject(
+        self.field ->
+        queryOp("$polygon", coords.toList)
+      )
+
+    def $box(lowerLeft: GeoCoords[_, _], upperRight: GeoCoords[_, _]) =
+      MongoDBObject(
+        self.field ->
+        queryOp("$box", MongoDBList(lowerLeft.toList, upperRight.toList)))
+
+    def $center[T: Numeric](center: GeoCoords[_, _], radius: T) =
+      MongoDBObject(
+        self.field ->
+        queryOp("$center", MongoDBList(center.toList, radius)))
+
+    def $centerSphere[T: Numeric](center: GeoCoords[_, _], radius: T) =
+      MongoDBObject(
+        self.field ->
+        queryOp("$centerSphere", MongoDBList(center.toList, radius)))
+  }
+
+}
+
+/**
+ *
+ * Trait to provide the \$intersects geospatial search method on appropriate callers
+ *
+ * @since 2.6.2
+ * @see http://www.mongodb.org/display/DOCS/Geospatial+Indexing
+ */
+trait GeoIntersectsOp extends QueryOperator {
+  private val oper = "$geoIntersects"
+  def $geoIntersects(geometry: DBObject) = queryOp(oper, geometry)
+}
+
+/**
+ *
  * Trait to provide the \$within geospatial search method on appropriate callers
  *
  *
@@ -501,10 +582,12 @@ trait GeoNearSphereOp extends QueryOperator {
  *   a) the order you specified your actual index in
  *   b) if you're using actual world maps or something else
  *
+ * Depreciated in MongoDB 2.4 use $geoWithin instead
+ *
  * @since 2.0
  * @see http://www.mongodb.org/display/DOCS/Geospatial+Indexing
  */
-trait GeoWithinOps extends QueryOperator {
+trait DeprecatedGeoWithinOps extends QueryOperator {
   self =>
   private val oper = "$within"